backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytesseract
Pillow
pydantic[email]
google-cloud-vision